fix and reorganize tests for custom emojis
